Link Each book widget to their own screen
Switched from displaying bookWidget in a column to displaying it in Listview.builder() :+1:  .

And also linked each of the listview item into their own respected screen with data dynamically ranging.

It looks bad, and only work for the main home page now, I will update the UI and apply it to all pages.