Implemented the Select Category page
User now will have an entirely different page for just picking the category of books that they want to sort there book out of.

Really simple and modern UI with nice background image, and it links to each specific category page and show only the books that exist in that user's inventory.