Fixed User input / Count Books
User now have limit of words when inputing stuff like the language for the books, the pages, etc.

The app will also auto count the amount of book and favourite everytime the homescreen or the myaccount page is fired, because it will ensure that the user data will always stay up to date.